Convert to use int64_t to aovid int oout-of-range
BUG=aomedia:518

+++ av1/encoder/mcomp.c
@@ -2348,7 +2348,7 @@ int av1_full_pixel_search(const AV1_COMP *cpi, MACROBLOCK *x, BLOCK_SIZE bsize,
 #define MVC(r, c)                                                         \
   (mvcost                                                                 \
        ? ((mvjcost[((r) != rr) * 2 + ((c) != rc)] + mvcost[0][((r)-rr)] + \
-           mvcost[1][((c)-rc)]) *                                         \
+           (int64_t)mvcost[1][((c)-rc)]) *                                \
               error_per_bit +                                             \
           4096) >>                                                        \
              13                                                           \
@@ -2938,7 +2938,7 @@ int av1_masked_full_pixel_diamond(const AV1_COMP *cpi, MACROBLOCK *x,
 #define MVC(r, c)                                                         \
   (mvcost                                                                 \
        ? ((mvjcost[((r) != rr) * 2 + ((c) != rc)] + mvcost[0][((r)-rr)] + \
-           mvcost[1][((c)-rc)]) *                                         \
+           (int64_t)mvcost[1][((c)-rc)]) *                                \
               error_per_bit +                                             \
           4096) >>                                                        \
              13                                                           \
